refactor(main): add main root
